Add initial codebase structure
This commit includes the initial setup of the project, including configuration files, component structure, and basic UI elements. The codebase is organized with components, pages, and utility functions to facilitate further development.
[skip gpt_engineer]
